xmltv: split the 947-line module into a focused package
xmltv.py had grown to a single 947-line XmltvGenerator class mixing channel
output, the 300-line per-programme writer, description enrichment, credits,
categories and media helpers. Split it into a gracenote2epg/xmltv/ package
where each concern is a short, focused mixin composed by the generator:

  generator.py    orchestration + header/footer (100 l)
  stations.py     <channel> elements
  programme.py    <programme> writer (DTD element order)
  descriptions.py description selection + enhanced info
  credits.py      DTD-ordered cast/crew
  categories.py   genres + EIT mapping
  media.py        ratings, icons, new/live

Method bodies are moved verbatim; the composed class preserves identical
behaviour. Verified byte-for-byte identical XMLTV output against the
pre-split code on the same cache and time window.
