+++ gracenote2epg/xmltv/__init__.py
@@ -1,0 +1,5 @@
+"""gracenote2epg.xmltv - XMLTV generation package."""
+
+from .generator import XmltvGenerator
+
+__all__ = ["XmltvGenerator"]
